# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Permissions:
    properties:
      product_module:
        type: boolean
      profile_module:
        type: boolean
      user_module:
        type: boolean
    type: object
...
  domain.Profile:
    properties:
      id:
        type: integer
      name:
        minLength: 4
        type: string
      permissions:
        $ref: '#/definitions/domain.Permissions'
    required:
    - name
    type: object
...
  dto.ItemsOutputDTO:
    properties:
      count:
        type: integer
      items: {}
    type: object
...
  dto.PermissionsInputDTO:
    properties:
      category_module:
        example: true
        type: boolean
      profile_module:
        example: true
        type: boolean
      user_module:
        example: true
        type: boolean
    type: object
...
  dto.ProfileInputDTO:
    properties:
      name:
        example: ADMIN
        type: string
      permissions:
        $ref: '#/definitions/dto.PermissionsInputDTO'
    type: object
...
  httphelper.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
info:
  contact:
    email: email@email.com
    name: Raul del Aguila
  description: Expenses API.
  title: Go - Expenses API
paths:
...
  /profile:
    get:
      consumes:
      - application/json
      description: Get profiles
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - example: 10
        in: query
        name: limit
        type: integer
      - example: descending order 'desc' or ascending order 'asc'
        in: query
        name: order
        type: string
      - example: 1
        in: query
        name: page
        type: integer
      - example: name
        in: query
        name: search
        type: string
      - example: '''updated_at'', ''created_at'', ''name'' or some other field of
          the response object'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ItemsOutputDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Get profiles
      tags:
      - Profile
    post:
      consumes:
      - application/json
      description: Insert profile
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: Profile model
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/dto.ProfileInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Insert profile
      tags:
      - Profile
  /profile/{id}:
    delete:
      consumes:
      - application/json
      description: Delete profile by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Delete profile
      tags:
      - Profile
    get:
      consumes:
      - application/json
      description: Get profile by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Get profile by ID
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Update profile by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      - description: Profile model
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/dto.ProfileInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Update profile
      tags:
      - Profile
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"